requirements.txt
setuptools>=78.1.0
openai-whisper==20231117
gTTS==2.4.0
python-dotenv==1.0.0
sounddevice==0.4.6
soundfile==0.12.1
numpy>=1.26.0
torch>=2.2.0
requests>=2.31.0
langdetect>=1.0.9
transformers>=4.36.0
ollama>=0.1.6
twilio>=8.12.0
python-dotenv>=1.0.0
flask>=3.0.0
waitress>=3.0.0 